LayoutTests/fast/scrolling/mac/smooth-scroll-iframe-expected.txt


PASS window.pageYOffset is 200
PASS successfullyParsed is true

TEST COMPLETE
